--HG--
branch : feature

+++ src/libs/vgeometry/vgobject.h
@@ -100,8 +100,8 @@ protected:
 private:
     QSharedDataPointer<VGObjectData> d;
 
-    static long double PerpDotProduct(const QPointF &p1, const QPointF &p2, const QPointF &t);
-    static long double GetEpsilon(const QPointF &p1, const QPointF &p2);
+    static double PerpDotProduct(const QPointF &p1, const QPointF &p2, const QPointF &t);
+    static double GetEpsilon(const QPointF &p1, const QPointF &p2);
 
     static int     PointInCircle (const QPointF &p, const QPointF &center, qreal radius);
 };
